change promo rate checking to double

+++ iwfpapp/lib/services/utilities/validators/promo_info_validator.dart
@@ -129,7 +129,7 @@ ValidationResponse isValidPromoRate(String promoRate) {
     response.messages.add(promoRateEmptyErrorMessage);
   }
   try {
-    int.parse(promoRate);
+    double.parse(promoRate);
   } catch (err) {
     response.valid = false;
     response.messages.add(promoRateNotIntErrorMessage);
